src/ml/requirements.txt
flask
xgboost
numpy
kafka-python
